requirement.txt
fastapi
uvicorn
Pillow
numpy
tensorflow